更改Cut API接口，返回第一帧时间戳
* int64_t Cut(int64_t start_ms, int64_t dur_ms, char *str,int str_size)
* 返回值为第一帧时间戳。调用后返回值<0，则为不可用; >=0,即为第一帧时间戳

+++ include/slice_video_proto.h
@@ -6,18 +6,18 @@ namespace HobotNebula {
 
 class Slice {
  public:
-  virtual int Cut(int64_t start_ms, int64_t dur_ms, char *str,
+  virtual int64_t Cut(int64_t start_ms, int64_t dur_ms, char *str,
                   int str_size) = 0;
 };
 
 class DMS_Slice : public Slice {
  public:
-  int Cut(int64_t start_ms, int64_t dur_ms, char *str, int str_size) override;
+  int64_t Cut(int64_t start_ms, int64_t dur_ms, char *str, int str_size) override;
 };
 
 class ADAS_Slice : public Slice {
  public:
-  int Cut(int64_t start_ms, int64_t dur_ms, char *str, int str_size) override;
+  int64_t Cut(int64_t start_ms, int64_t dur_ms, char *str, int str_size) override;
 };
 }
